requirements.txt
click==6.7
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
scikit-learn==0.19.1
scipy==1.0.0
Werkzeug==0.13
